Added functionality to delete notes on the web app.
I added a delete button in the HTML for each note. Next, I added the necessary Javascript to hit the DeleteNote method that I added in SourceExplorer. Finally, I updated the Javascript to add the delete note button automatically when a new note is created and to refresh the page.