Completed AddSource Implementation
The MainPage now shows the AddSource form as a dialog of sorts when the Add button is clicked. The MainPageForm also passes its controller's authservice to allow AddSourceForm's controller to have access to the logged in user. Furthermore, the authservice now has implementation for adding source through the web access route.