Fixed Add Source Bug
The errant MessageBox line is now correctly encapsulated by an else statement so it is thrown only when necessary. The MainPageController now queries the sources from the db instead of maintaining a local copy which removes the issue of the local copy getting outdated. A new ComboBox was added to the AddSource dialog to facilitate functionality for when sources of different types are implemented on Desktop.