Updated the Desktop UI to utilize a new Note custom user control.
I created a new custom user control for notes. I added a text box for the note text, a listview for the tags, along with buttons. The buttons are to save changes to note, delete note, add new tag, or delete tag. I also added the NoteControl to the source control and added the necessary events. Finally, I have the note text displayed and added message boxes for the button clicks to test. Waiting on group members to confirm design before I continue.