Implemented the Search Tags functionality.
I finished up the backend functionality of the search tags feature. I added the SearchTag route in SourceExplorer in the web app. Added the GetSourcesWithMatchingTag method in the AuthService class. And I added the getSourcesWithMatchingNoteTags class in the MainPageController class. I also cleaned up a few things and added some documentation.