Switched Framework for Video Player
The old video player framework lacked some features requested and was replaced by a fully built component in the Windows Media Player. The WMP now handles all the functionality related to playback without the system having to worry about it. As a consequence, videos now need to be downloaded for playback.