Adjusted AddTags implementation
Previously the AddTags function didn't check if a tag already existed in the db before adding it, this has been rectified to make the implementation more consistent. Also included here is an update to Tags being displayed on existing notes in preparation for adding and deleting tags on existing notes. The listview's items now have checkboxes so that the user can select tags to be deleted.