Started working on the Notes UI update.
First, I removed the necessary controls from the SourceExplorer form. Next, I added a "Add Note" button and a Notes listview. Next, I updated the AddNoteForm that we had, but was not in use. I then added the functionality to show the window when the AddNote button is clicked. Currently the add note functionality is working, but the notes are not being refreshed when the note is added. I have also created a new NoteForm, which will be displayed when a note listview item is clicked.