Worked more on the notes UI update on the desktop.
I removed the unneeded NoteControl & AddNoteControl classes. I added some docs, and removed some unneeded code in the SourceForm, NoteForm, and AddNoteForm classes. Finally, I fixed the windows so that they all open centered on top of the parent window. Currently, the functionality appears to be working. Still have a couple of bugs to address.