Updated the desktop search function.
I changed the tags list view to a checked list box. If a box is checked, that item is removed. The search is adjusted as items are added or removed. I also disabled the Delete Tag button on the NoteForm. I also added an event to enable the button if a tag is selected.