Found and resolved the bug when clicking the "Go Back" button on the web.
I found that the bug only occured when I would search a tag, then remove the search tag, then opening a source, and clicking go back. I found the issue to be in regard to the browser history after those chagnes. I added a method to update the history after a tag change, and the issue appears to have been resolved.