Updated the Add Source Form to correct error messages.
Fixed the error messages to only display one message regardless of the error. If all fields are blank, an error message for all fields is displayed. If the title is blank, an error message regarding the title is displayed. If the source is not uploaded, an error message regarding the missing source is displayed.